+++ mayan/apps/acls/tests/mixins.py
@@ -1,12 +1,17 @@
 from __future__ import unicode_literals
 
+from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ImproperlyConfigured
 
-from mayan.apps.acls.models import AccessControlList
+from mayan.apps.common.tests.mixins import TestModelTestMixin
 from mayan.apps.permissions.tests.mixins import (
     PermissionTestMixin, RoleTestCaseMixin, RoleTestMixin
 )
 from mayan.apps.user_management.tests.mixins import UserTestCaseMixin
+
+from ..classes import ModelPermission
+from ..models import AccessControlList
+from ..permissions import permission_acl_edit, permission_acl_view
 
 
 class ACLTestCaseMixin(RoleTestCaseMixin, UserTestCaseMixin):
@@ -27,7 +32,7 @@ class ACLTestCaseMixin(RoleTestCaseMixin, UserTestCaseMixin):
         )
 
 
-class ACLTestMixin(PermissionTestMixin, RoleTestMixin):
+class ACLTestMixin(PermissionTestMixin, RoleTestMixin, TestModelTestMixin):
     auto_create_test_role = True
 
     def _create_test_acl(self):
@@ -39,3 +44,32 @@ class ACLTestMixin(PermissionTestMixin, RoleTestMixin):
         super(ACLTestMixin, self).setUp()
         if self.auto_create_test_role:
             self._create_test_role()
+
+    def _inject_test_object_content_type(self):
+        self.test_object_content_type = ContentType.objects.get_for_model(
+            model=self.test_object
+        )
+
+        self.test_content_object_view_kwargs = {
+            'app_label': self.test_object_content_type.app_label,
+            'model_name': self.test_object_content_type.model,
+            'object_id': self.test_object.pk
+        }
+
+    def _setup_test_object(self):
+        self._create_test_model()
+        self._create_test_object()
+        ModelPermission.register(
+            model=self.test_object._meta.model, permissions=(
+                permission_acl_edit, permission_acl_view,
+            )
+        )
+
+        self._create_test_permission()
+        ModelPermission.register(
+            model=self.test_object._meta.model, permissions=(
+                self.test_permission,
+            )
+        )
+
+        self._inject_test_object_content_type()